SRC/Amail/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-appconf==1.0.5
django-select2==7.10.0
django-utils-six==2.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
djangorestframework~=3.13.1